# FSA.Tickets-Please/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SIMS incident/signals management API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Incidents/Status/{id}/{statusId}": {
      "post": {
        "tags": [
          "Incidents"
        ],
        "summary": "Update status of an incident",
        "operationId": "UpdateStatus",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "statusId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BaseIncident"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
    "/api/Incidents/LeadOfficer": {
      "post": {
        "tags": [
          "Incidents"
        ],
        "summary": "Update incident(s) lead officer",
        "operationId": "UpdateLeadOfficer",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateLeadOfficerModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
    "/api/Signals/{id}": {
      "get": {
        "tags": [
          "Signals"
        ],
        "summary": "Get Signal by id",
        "operationId": "GetSignal",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SimsSignal"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Signals": {
      "put": {
        "tags": [
          "Signals"
        ],
        "summary": "Replace an Signal",
        "operationId": "UpdateSignal",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SimsSignal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SimsSignal"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Signals"
        ],
        "summary": "Create an Signal",
        "operationId": "CreateSignal",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SimsSignal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SimsSignal"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Signals/Status/{id}/{statusId}": {
      "post": {
        "tags": [
          "Signals"
        ],
        "summary": "Update status of an Signal",
        "operationId": "UpdateSignalStatus",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "statusId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SimsSignal"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Signals/LeadOfficer": {
      "post": {
        "tags": [
          "Signals"
        ],
        "summary": "Update Signal(s) lead officer",
        "operationId": "UpdateSignalLeadOfficer",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateLeadOfficerModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Signals/Dashboard": {
      "post": {
        "tags": [
          "Signals"
        ],
        "summary": "Signals dashboard search",
        "operationId": "GetSignalDashboard",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DashboardSearchViewModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SignalDashboardItem"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Signals/Close/NoIncident": {
      "post": {
        "tags": [
          "Signals"
        ],
        "summary": "Close no incident",
        "operationId": "CloseNoIncident",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SimsSignalCloseNoIncident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Signals/Close/Create": {
      "post": {
        "tags": [
          "Signals"
        ],
        "summary": "Close create incident",
        "operationId": "CloseCreateIncident",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SimsSignalCloseCreateIncident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/IncidentIdViewModel"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Signals/Close/Link": {
      "post": {
        "tags": [
          "Signals"
        ],
        "summary": "Close link incident",
        "operationId": "CloseLinkIncident",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SimsSignalCloseLinkIncident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Stakeholders/{incidentSignal}/{id}": {
      "get": {
        "tags": [
          "Stakeholders"
        ],
        "summary": "Get All Stakeholders",
        "operationId": "GetStakeholder",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentSignal",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SimsStakeholder"
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Stakeholders"
        ],
        "summary": "Remove a stakeholder to an incident",
        "operationId": "RemoveStakeholder",
        "parameters": [
          {
            "in": "path",
            "name": "incidentSignal",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Stakeholders/{incidentSignal}": {
      "post": {
        "tags": [
          "Stakeholders"
        ],
        "summary": "Add a new stakeholder to an incident",
        "operationId": "AddStakeholder",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentSignal",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SimsStakeholder"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SimsStakeholder"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Stakeholders"
        ],
        "summary": "Update a stakeholder",
        "operationId": "UpdateStakeholder",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentSignal",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SimsStakeholder"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SimsStakeholder"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BaseIncident": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "commonId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "mostUniqueId": {
          "format": "uuid",
          "type": "string",
          "readOnly": true
        },
        "incidentTitle": {
          "type": "string",
          "readOnly": true
        },
        "incidentTypeId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "contactMethodId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "statusId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "signalStatusId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "notifierId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "principalFBOId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "priorityId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "classificationId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "dataSourceId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "signalUrl": {
          "type": "string",
          "readOnly": true
        },
        "productTypeId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "leadOfficer": {
          "type": "string",
          "readOnly": true
        },
        "leadOffice": {
          "type": "string",
          "readOnly": true
        },
        "adminLeadId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "fieldOfficer": {
          "type": "string",
          "readOnly": true
        },
        "leadLocalAuthorityId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "laAdvised": {
          "type": "boolean",
          "readOnly": true
        },
        "deathIllnessId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "receivedOn": {
          "format": "date-time",
          "type": "string",
          "readOnly": true
        },
        "incidentCreated": {
          "format": "date-time",
          "type": "string",
          "readOnly": true
        },
        "incidentClosed": {
          "format": "date-time",
          "type": "string",
          "readOnly": true
        },
        "lastChangedBy": {
          "type": "string",
          "readOnly": true
        },
        "lastChangedDate": {
          "format": "date-time",
          "type": "string",
          "readOnly": true
        }
      }
    },
...
    "UpdateLeadOfficerModel": {
      "required": [
        "ids",
        "officer"
      ],
      "type": "object",
      "properties": {
        "ids": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        },
        "officer": {
          "type": "string"
        }
      }
    },
    "DashboardSearchViewModel": {
      "type": "object",
      "properties": {
        "search": {
          "type": "string"
        },
        "pageSize": {
          "format": "int32",
          "type": "integer"
        },
        "pageNo": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "SimsSignal": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "commonId": {
          "format": "int32",
          "type": "integer"
        },
        "sptId": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "signalStatusId": {
          "format": "int32",
          "type": "integer"
        },
        "priority": {
          "type": "string"
        },
        "leadOfficer": {
          "type": "string"
        },
        "alertType": {
          "type": "string"
        },
        "countryOfOrigin": {
          "type": "string"
        },
        "notifyingCountry": {
          "type": "string"
        },
        "isEu": {
          "type": "boolean"
        },
        "foodOrFeed": {
          "type": "string"
        },
        "baseProduct": {
          "type": "string"
        },
        "manufacturer": {
          "type": "string"
        },
        "hazard": {
          "type": "string"
        },
        "hazardGroup": {
          "type": "string"
        },
        "dataSource": {
          "type": "string"
        },
        "sourceType": {
          "type": "string"
        },
        "publishedDate": {
          "format": "date-time",
          "type": "string"
        },
        "insertedDate": {
          "format": "date-time",
          "type": "string"
        },
        "sourceLink": {
          "type": "string"
        },
        "lastUpdated": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "SignalDashboardItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "commonId": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "priority": {
          "type": "string"
        },
        "leadOfficer": {
          "type": "string"
        },
        "countryOfOrigin": {
          "type": "string"
        },
        "hazard": {
          "type": "string"
        },
        "baseProduct": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "dateReceived": {
          "format": "date-time",
          "type": "string"
        },
        "updated": {
          "format": "date-time",
          "type": "string"
        },
        "links": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        }
      }
    },
    "SimsSignalCloseNoIncident": {
      "type": "object",
      "properties": {
        "signalId": {
          "format": "int32",
          "type": "integer"
        },
        "teamId": {
          "format": "int32",
          "type": "integer"
        },
        "reasonId": {
          "format": "int32",
          "type": "integer"
        },
        "statusCloseId": {
          "format": "int32",
          "type": "integer"
        },
        "userReason": {
          "type": "string"
        }
      }
    },
    "SimsSignalCloseCreateIncident": {
      "type": "object",
      "properties": {
        "reasonNote": {
          "type": "string"
        },
        "signalId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "IncidentIdViewModel": {
      "type": "object",
      "properties": {
        "incidentId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "SimsSignalCloseLinkIncident": {
      "type": "object",
      "properties": {
        "signalId": {
          "format": "int32",
          "type": "integer"
        },
        "incidentId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "SimsStakeholder": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "hostId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "govDept": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "discriminatorId": {
          "format": "int32",
          "type": "integer"
        },
        "addressId": {
          "format": "int32",
          "type": "integer"
        },
        "addressTitle": {
          "type": "string"
        }
      }
    }
  }
}